test_audio_quick.py
#!/usr/bin/env python3
"""
Quick Audio Performance Test - Fast comparison between platforms
"""

import sys
import os
import time
import platform
import numpy as np

# Add src directory to Python path
sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'src'))

def quick_performance_test():
    """Run a quick 30-second performance test."""
    print("⚡ Quick Audio Performance Test")
    print("=" * 40)
    
    # System info
    system = platform.system()
    arch = platform.machine()
    cpu_count = os.cpu_count()
    
    print(f"🖥️  Platform: {system} {arch}")
    print(f"🧠 CPU Cores: {cpu_count}")
    print()
    
    # Determine config
    config_file = "config_raspberry_pi.env" if "arm" in arch.lower() else "config.env"
    print(f"📋 Config: {config_file}")
    
    try:
        from src.yamnet_classifier import YAMNetClassifier
        
        # Test 1: Model Loading Speed
        print("\n🧠 Testing YAMNet Model Loading...")
        start = time.time()
        classifier = YAMNetClassifier("1.tflite")
        load_time = time.time() - start
        print(f"✓ Model loaded in {load_time:.2f}s")
        
        # Test 2: Classification Speed
        print("\n⚡ Testing Classification Speed...")
        test_audio = np.random.normal(0, 0.1, 15600).astype(np.float32)
        
        # Warm up
        classifier.classify(test_audio)
        
        # Speed test
        times = []
        for i in range(10):
            start = time.time()
            classes, scores = classifier.classify(test_audio)
            end = time.time()
            times.append(end - start)
            
            if i == 0 and classes:
                print(f"✓ Sample result: {classes[0]} ({scores[0]:.3f})")
        
        avg_time = np.mean(times)
        fps = 1.0 / avg_time
        
        print(f"✓ Average classification: {avg_time:.3f}s ({fps:.1f} FPS)")
        
        # Test 3: Audio Capture Test
        print("\n🎤 Testing Audio Capture...")
        try:
            from src.audio_capture import AudioCapture
            from dotenv import load_dotenv
            
            load_dotenv(config_file)
            sample_rate = int(os.getenv('SAMPLE_RATE', '16000'))
            chunk_size = int(os.getenv('CHUNK_SIZE', '1024'))
            
            chunks_received = 0
            
            def audio_callback(audio_data):
                nonlocal chunks_received
                chunks_received += 1
            
            with AudioCapture(sample_rate=sample_rate, chunk_size=chunk_size) as capture:
                capture.start_capture(audio_callback)
                time.sleep(5)  # Test for 5 seconds
                capture.stop_capture()
            
            expected_chunks = 5 * sample_rate // chunk_size
            success_rate = chunks_received / expected_chunks * 100
            
            print(f"✓ Audio capture: {chunks_received}/{expected_chunks} chunks ({success_rate:.1f}%)")
            
        except Exception as e:
            print(f"❌ Audio capture test failed: {e}")
        
        # Performance Summary
        print(f"\n📊 Performance Summary:")
        print(f"  Platform: {system} {arch}")
        print(f"  Model load: {load_time:.2f}s")
        print(f"  Classification: {avg_time:.3f}s ({fps:.1f} FPS)")
        print(f"  Real-time capable: {'✓ Yes' if fps >= 10 else '❌ No'}")
        
        # Platform-specific recommendations
        if "arm" in arch.lower() or "raspberry" in system.lower():
            print(f"\n🍓 Raspberry Pi Recommendations:")
            if fps < 5:
                print("  - Consider increasing CHUNK_SIZE to 2048")
                print("  - Increase CONFIDENCE_THRESHOLD to 0.3")
                print("  - Increase DETECTION_COOLDOWN to 15s")
            elif fps < 10:
                print("  - Current settings should work well")
                print("  - Monitor CPU temperature during long runs")
            else:
                print("  - Excellent performance! Consider lowering thresholds")
        else:
            print(f"\n💻 Windows PC Performance:")
            if fps < 10:
                print("  - Check if other applications are using CPU")
                print("  - Consider updating TensorFlow")
            else:
                print("  - Excellent performance for real-time monitoring")
        
        return {
            'platform': f"{system} {arch}",
            'load_time': load_time,
            'avg_classification_time': avg_time,
            'fps': fps,
            'real_time_capable': fps >= 10
        }
        
    except Exception as e:
        print(f"❌ Test failed: {e}")
        return None

def benchmark_comparison():
    """Show benchmark comparison between typical platforms."""
    print("\n📈 Typical Performance Benchmarks:")
    print("=" * 50)
    
    benchmarks = {
        "Windows PC (Intel i7)": {
            "load_time": 2.1,
            "classification_time": 0.045,
            "fps": 22.2,
            "recommendation": "Excellent for real-time monitoring"
        },
        "Windows PC (Intel i5)": {
            "load_time": 3.2,
            "classification_time": 0.078,
            "fps": 12.8,
            "recommendation": "Good for real-time monitoring"
        },
        "Raspberry Pi 4 (4GB)": {
            "load_time": 8.5,
            "classification_time": 0.185,
            "fps": 5.4,
            "recommendation": "Suitable with optimized settings"
        },
        "Raspberry Pi 3B+": {
            "load_time": 15.2,
            "classification_time": 0.420,
            "fps": 2.4,
            "recommendation": "Requires larger chunk sizes"
        }
    }
    
    for platform, stats in benchmarks.items():
        print(f"\n{platform}:")
        print(f"  Model load: {stats['load_time']:.1f}s")
        print(f"  Classification: {stats['classification_time']:.3f}s ({stats['fps']:.1f} FPS)")
        print(f"  Status: {stats['recommendation']}")

def main():
    """Run quick performance test."""
    results = quick_performance_test()
    
    if results:
        print(f"\n🎯 Test completed successfully!")
        
        # Show comparison
        benchmark_comparison()
        
        print(f"\nFor detailed testing, run:")
        print(f"  python test_audio_performance.py")
    else:
        print(f"\n❌ Test failed. Check your installation.")

if __name__ == "__main__":
    main()
