requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
pyaudio>=0.2.11
python-telegram-bot>=20.0
pydub>=0.25.1
scipy>=1.7.0
python-dotenv>=0.19.0
colorlog>=6.6.0
